Moved DTOs structs to a separated file
