+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/BackpackEditorHUDV2/VRMExport/NFTItemDTOs.cs
@@ -1,0 +1,34 @@
+using System;
+
+namespace DCL.Backpack
+{
+    [Serializable]
+    public struct NFTItemsDTO
+    {
+        public NFTDataDTO[] data;
+    }
+
+    [Serializable]
+    public struct NFTDataDTO
+    {
+        public string name;
+        public string thumbnail;
+        public WearableDataDTO data;
+        public string creator; // Creator wallet
+        public string creatorName;
+        public string creatorImageUrl;
+        public string urn;
+    }
+
+    [Serializable]
+    public struct WearableDataDTO
+    {
+        public WearableDTO wearable;
+    }
+
+    [Serializable]
+    public struct WearableDTO
+    {
+        public string category;
+    }
+}
