+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/BackpackEditorHUDV2/VRMExport/NFTItemDTOs.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 5b81d231be5147b69614bfba5117f752
+timeCreated: 1715186818
